+++ src/Control/Distributed/Raketka/NewServerInfo.hs
@@ -18,15 +18,16 @@ newServerInfo::Content tag ps s c =>
     -> Process ()
 
 newServerInfo s1@(Tagged server0@Server{..}) ping0 pid0 = do
-    liftIO $ printf "%s received %s from %s\n" spid (show ping0) pid0
-    join $ liftIO $ atomically $ do
-        old_pids1 <- readTVar servers
-        let old_pids2 = peer_pids old_pids1
-        writeTVar servers $ onPeerConnected' old_pids1 pid0
+    say $ printf "%s received %s from %s\n" spid (show ping0) pid0
+    old_pids1 <- la $ readTVar servers
+    let old_pids2 = peer_pids old_pids1
+    la $ writeTVar servers $ onPeerConnected' old_pids1 pid0
 
-        when (ping0 == Ping) $ sendRemote s1 pid0 $ Info Pong spid 
-    
-        -- monitor the new server
-        pure (when (pid0 `notElem` old_pids2) $ void $ monitor pid0)
+    if (ping0 == Ping) then 
+        la (sendRemote s1 pid0 $ Info Pong spid)
+        else pure ()
+
+    -- monitor the new server
+    when (pid0 `notElem` old_pids2) $ void $ monitor pid0
 
     onPeerConnected s1 pid0